Security and robustness hardening
Multiple security fixes and robustness improvements across the codebase:

- Add small deps (async-trait, time) used by middleware/handlers.
- Config: new rate_limit_gets option to allow separate GET rate limits.
- Middleware: rate limiter now applies to GET and POST (separate buckets) and uses constant-time CSRF comparison; added ClientIp extractor that respects proxy headers.
- Admin: brute-force lockout per-IP (hashed keys) with opportunistic cleanup; reset on success; improved login cookie handling (Max-Age + Secure derived from config); stricter redirect validation (reject backslash/%5C).
- Zip upload/restore: replace unbounded std::io::copy with copy_limited and a 256 MiB per-entry cap to mitigate zip-bomb decompression.
- Router: add per-route DefaultBodyLimit to several endpoints and set security headers (CSP, HSTS, Permissions-Policy).
- Request tracking: store SHA-256(IP) hashes instead of raw IPs for ACTIVE_IPS and show 8-char prefixes in diagnostics.
- DB pool: add connection_timeout to avoid blocking spawn_blocking threads when pool is exhausted.
- Workers: add exponential back-off with jitter for repeated errors to reduce DB storming; reset counter on success or empty queue.
- Multipart/poll handling: server-side caps on poll question length, option count and each option length.
- Files: fix free_bytes calculation by casting to u64 to avoid overflow on large values.

These changes aim to reduce attack surface (XSS/DoS/brute-force/zip bombs), prevent retention of plaintext IPs in memory, and improve stability under load.

+++ src/utils/files.rs
@@ -264,7 +264,7 @@ pub fn save_upload(
             if let Ok(path_cstr) = std::ffi::CString::new(dir_bytes.as_bytes()) {
                 let mut stat: libc::statvfs = std::mem::zeroed();
                 if libc::statvfs(path_cstr.as_ptr(), &mut stat) == 0 {
-                    let free_bytes = stat.f_bavail * stat.f_frsize;
+                    let free_bytes = u64::from(stat.f_bavail) * stat.f_frsize;
                     let needed = (final_data.len() as u64).saturating_mul(2);
                     if free_bytes < needed {
                         return Err(anyhow::anyhow!(
